Added photo of hotlines, and barangay officials as well as some revisions

+++ app/src/main/java/com/example/e_reklamo/PostGetAdapter.kt
@@ -1,5 +1,6 @@
 package com.example.e_reklamo
 
+import android.annotation.SuppressLint
 import android.app.Dialog
 import android.content.Context
 import android.text.format.DateUtils
@@ -8,11 +9,13 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
+import android.widget.LinearLayout
 import android.widget.ProgressBar
 import android.widget.TextView
 import android.widget.Toast
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.firebase.database.FirebaseDatabase
 import io.github.jan.supabase.createSupabaseClient
 import io.github.jan.supabase.storage.Storage
@@ -40,6 +43,10 @@ class PostGetAdapter(private val posts: List<Map<String, Any>>, private val user
         val deletePost: TextView = view.findViewById(R.id.deletepost)
         val postImageView: ImageView = view.findViewById(R.id.postImage)
         val postProfileImage: ImageView = view.findViewById(R.id.postprofile)
+        val statusLayout: LinearLayout = view.findViewById(R.id.statuslayout)
+        val postStatus: TextView = view.findViewById(R.id.status)
+        val postRead: TextView = view.findViewById(R.id.read)
+
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): PostViewHolder {
@@ -47,11 +54,11 @@ class PostGetAdapter(private val posts: List<Map<String, Any>>, private val user
         return PostViewHolder(view)
     }
 
+    @SuppressLint("SetTextI18n")
     override fun onBindViewHolder(holder: PostViewHolder, position: Int) {
         val post = posts[position]
 
-        // Extract and bind the data to the views
-        holder.accountName.text = post["name"] as String
+        holder.accountName.text = "User${(post["keysecret"] as String).take(5)}"
         holder.accountPosition.text = post["position"] as String
         holder.postContent.text = post["content"] as String
         Glide.with(holder.itemView.context)
@@ -65,10 +72,6 @@ class PostGetAdapter(private val posts: List<Map<String, Any>>, private val user
         holder.postImageView.setOnClickListener {
             showImageDialog(holder.itemView.context, post["imageUrl"] as String)
         }
-        Glide.with(holder.itemView.context)
-            .load(post["profileImage"] as String)
-            .circleCrop()
-            .into(holder.postProfileImage)
 
         // Get the timestamp from the post data (saved as long)
         val timestamp = post["timestamp"] as Long
@@ -78,7 +81,24 @@ class PostGetAdapter(private val posts: List<Map<String, Any>>, private val user
         holder.postDate.text = timeAgo
         val PostAccountTyper = post["keysecret"] as String
 
+        val postStatus = post["status"] as String
+        val postRead = post["read"] as String
+        holder.postStatus.text = "Status: $postStatus"
+        if(postRead == "yes"){
+            holder.postRead.text = "Seen"
+            holder.postRead.setCompoundDrawablesWithIntrinsicBounds(0, 0, R.drawable.round_check_circle_241, 0)
+        } else {
+            holder.postRead.text = "Sent"
+            holder.postRead.setCompoundDrawablesWithIntrinsicBounds(0, 0, R.drawable.round_check_circle_outline_24, 0)
+        }
         holder.deletePost.visibility = if(userKey == PostAccountTyper) View.VISIBLE else View.GONE
+        holder.statusLayout.visibility = if(userKey == PostAccountTyper) View.VISIBLE else View.GONE
+        if(userKey == PostAccountTyper){
+            Glide.with(holder.itemView.context)
+                .load(post["profileImage"] as String)
+                .circleCrop()
+                .into(holder.postProfileImage)
+        }
         holder.deletePost.setOnClickListener {
             deletePost(post["postkey"].toString(), holder.itemView.context, post["keysecret"].toString(), post["imageUrl"].toString())
         }
@@ -123,33 +143,45 @@ class PostGetAdapter(private val posts: List<Map<String, Any>>, private val user
         Log.d("ytr", "$database")
 
         // Delete from Firebase
-        database.removeValue()
-            .addOnSuccessListener {
-                // Delete from Supabase
-                val storage = supabase.storage.from("images")
-                val path = imageUrl.substringAfterLast("/") // Extract theimage path from the URL
+        MaterialAlertDialogBuilder(context)
+            .setTitle("Confirm Delete").setMessage("Are you sure you want to delete this post?")
+            .setPositiveButton("Delete") { dialog, which ->
+                // User clicked Delete, proceed with deletion
+                Progreessbar.visibility = View.VISIBLE
+                val database = FirebaseDatabase.getInstance().getReference("Users/$userId/Post/$postKey")
+                Log.d("ytr", "$database")
 
-                runBlocking {
-                    launch {
-                        try {
-                            val response = storage.delete(listOf(path)) // Use await() to get the result
-                            Toast.makeText(context, "Post and image deleted successfully", Toast.LENGTH_SHORT).show()
-                        } catch (error: Exception) {
-                            Toast.makeText(context, "Error deleting image from Supabase: ${error.message}", Toast.LENGTH_SHORT).show()
+                // Delete from Firebase
+                database.removeValue()
+                    .addOnSuccessListener {
+                        // Delete from Supabase
+                        val storage = supabase.storage.from("images")
+                        val path =imageUrl.substringAfterLast("/") // Extract the image path from the URL
+
+                        runBlocking {
+                            launch {
+                                try {
+                                    val response = storage.delete(listOf(path)) // Use await() to get the result
+                                    Toast.makeText(context, "Post deleted successfully", Toast.LENGTH_SHORT).show()
+                                } catch (error: Exception) {
+                                    Toast.makeText(context, "Error deleting image from Supabase: ${error.message}", Toast.LENGTH_SHORT).show()
+                                }
+                            }
                         }
+                        Progreessbar.visibility = View.GONE
                     }
-                }
-                Progreessbar.visibility = View.GONE
-            }
-            .addOnFailureListener { error ->
-                Toast.makeText(context, "Error deleting post: ${error.message}", Toast.LENGTH_SHORT).show()
-            }
+                    .addOnFailureListener { error ->
+                        Toast.makeText(context, "Error deleting post: ${error.message}", Toast.LENGTH_SHORT).show()
+                    }
 
-        val index = posts.indexOfFirst { it["postkey"] == postKey }
-        if (index != -1) {
-            (posts as MutableList).removeAt(index)
-            notifyItemRemoved(index)
-        }
+                val index = posts.indexOfFirst { it["postkey"] == postKey }
+                if (index != -1) {
+                    (posts as MutableList).removeAt(index)
+                    notifyItemRemoved(index)
+                }
+            }
+            .setNegativeButton("Cancel", null) // Do nothing on Cancel
+            .show()
     }
     //endregion
 }
